+++ packages/transaction-controller/src/utils/feature-flags.test.ts
@@ -74,6 +74,8 @@ describe('Feature Flags Utils', () => {
     getFeatureFlagsMock.mockReturnValue({
       cacheTimestamp: 0,
       remoteFeatureFlags: featureFlags,
+      rawRemoteFeatureFlags: {},
+      localOverrides: {},
     });
   }